Make backup a no-op when file doesn't exist.
Previously it would throw an error.

+++ src/python/materiality/util/fileutil.py
@@ -17,9 +17,12 @@ def umask(mask):
 
 @contextmanager
 def backup(path):
-  bak = path + '.bak'
-  os.rename(path, bak)
-  yield bak  # Caller decides when to delete bak.
+  if os.path.exists(path):
+    bak = path + '.bak'
+    os.rename(path, bak)
+    yield bak  # Caller decides when to delete bak.
+  else:
+    yield None
 
 
 @contextmanager
